Add files via upload
Conversation with the supplier at ALI BABA is not easy work. I faced a lot of problems regarding quality, price (Landing Cost), and profit margin. I chose the supplier with trademark and trade assurance. Shipping at time is one of the crucial in sourcing. But I managed it on time.
I have discussed this with many suppliers for more than 15. I selected 5 and ordered a sample of them. I picked one of the best with the quality, price, and profit margin point of view. My supplier gave me 1.37$ per unit with a packaging box (color box). The profit margin is more than 40 %.